pcr_quic::context

Struct PcrCryptoContext

pub struct PcrCryptoContext { /* private fields */ }

PCR Crypto Context for managing epoch keys and rekeying

Implementations

impl PcrCryptoContext

pub fn new( is_server: bool, params: NegotiatedPcrParams, initial_secret: &[u8; 32], connection_id: Vec<u8>, ) -> Result<Self>

Create a new PCR crypto context after handshake

Arguments

is_server - Whether we are the server

params - Negotiated PCR parameters

initial_secret - Initial shared secret from TLS exporter

connection_id - Connection ID for key derivation

pub fn kem_public_key(&self) -> &[u8]

Get our KEM public key (to send to peer)

pub fn set_peer_kem_public_key(&mut self, pk: Vec<u8>) -> Result<()>

Set peer’s KEM public key (received from peer)

pub fn current_epoch(&self) -> Epoch

Get current epoch number

pub fn next_epoch(&self) -> Option<Epoch>

Get next epoch number (if in transition)

pub fn should_rekey(&self) -> bool

Check if rekey is due

pub fn initiate_rekey(&mut self) -> Result<PcrRekeyFrame>

Initiate a rekey (sender side)

Returns the PCR_REKEY frame to send and the new shared secret

pub fn process_rekey_frame(&mut self, frame: &PcrRekeyFrame) -> Result<()>

Process received PCR_REKEY frame (receiver side)

pub fn activate_next_epoch(&mut self) -> Result<()>

Activate the next epoch (switch to new keys)

pub fn retire_epoch(&mut self, epoch: Epoch)

Retire old epoch keys

pub fn should_retire_epoch(&self, epoch: Epoch, pto: Option<Duration>) -> bool

Check if an epoch should be retired

pub fn send_key(&mut self) -> Option<&mut PcrPacketKey>

Get send key for current epoch

pub fn send_key_for_epoch(&mut self, epoch: Epoch) -> Option<&mut PcrPacketKey>

Get send key for a specific epoch

pub fn recv_key_for_epoch(&mut self, epoch: Epoch) -> Option<&mut PcrPacketKey>

Get receive key, trying current then next epoch

pub fn try_decrypt( &mut self, pn: u64, cid: &[u8], ad: &[u8], ciphertext: &[u8], ) -> Option<(Vec<u8>, Epoch)>

Try to decrypt with available epochs

pub fn encrypt( &mut self, pn: u64, cid: &[u8], ad: &[u8], plaintext: &[u8], ) -> Result<Vec<u8>>

Encrypt a packet

pub fn params(&self) -> &NegotiatedPcrParams

Get negotiated parameters